+++ src/Entities/Order.php
@@ -15,24 +15,15 @@ use GingerPluginSdk\Properties\Amount;
 use GingerPluginSdk\Properties\Currency;
 use GingerPluginSdk\Properties\Status;
 
-class Order implements MultiFieldsEntityInterface
+final class Order implements MultiFieldsEntityInterface
 {
     use HelperTrait;
     use MultiFieldsEntityTrait;
     use FieldsValidatorTrait;
     use SingleFieldTrait;
 
-    private BaseField $merchantOrderId;
-    private BaseField $webhookUrl;
-    private BaseField $returnUrl;
-    private BaseField $description;
-    private BaseField $created;
-    private BaseField $id;
-    private BaseField $lastTransactionAdded;
-    private BaseField $merchantId;
-    private BaseField $modified;
-    private BaseField $projectId;
-    private BaseField $completed;
+    private BaseField|null $id = null;
+    private BaseField|null $merchantOrderId = null;
 
     public function __construct(
         private Currency     $currency,
@@ -45,6 +36,7 @@ class Order implements MultiFieldsEntityInterface
         private ?Flags       $flags = null,
         ?string              $id = null,
         private ?Status      $status = null,
+        ?string              $merchantOrderId = null,
         mixed                ...$additionalProperties
     )
     {
@@ -53,32 +45,37 @@ class Order implements MultiFieldsEntityInterface
             value: $id
         );
 
+        if ($merchantOrderId) $this->merchantOrderId = $this->createSimpleField(
+            propertyName: 'merchantOrderId',
+            value: $merchantOrderId
+        );
+
         if ($additionalProperties) $this->filterAdditionalProperties($additionalProperties);
     }
 
-    public function getId(): string|null
+    public function getId(): BaseField|null
     {
-        return isset($this->id) ? $this->id->get() : null;
+        return $this->id;
     }
 
-    public function getStatus(): string|null
+    public function getStatus(): Status|null
     {
-        return $this->status?->get();
+        return $this->status;
     }
 
-    public function getClient(): array
+    public function getClient(): Client
     {
-        return $this->client?->toArray();
+        return $this->client;
     }
 
-    public function getCustomer(): array
+    public function getCustomer(): Customer
     {
-        return $this->customer->toArray();
+        return $this->customer;
     }
 
-    public function getAmount(): int
+    public function getAmount(): Amount
     {
-        return $this->amount->get();
+        return $this->amount;
     }
 
     public function getOrderLines(): ?OrderLines
@@ -86,29 +83,14 @@ class Order implements MultiFieldsEntityInterface
         return $this->orderLines;
     }
 
-    public function getReturnUrl(): string
+    public function getMerchantOrderId(): BaseField|null
     {
-        return $this->returnUrl->get();
+        return $this->merchantOrderId;
     }
 
-    public function getWebhookUrl(): string
+    public function getExtra(): Extra
     {
-        return $this->webhookUrl->get();
-    }
-
-    public function getMerchantOrderId(): string
-    {
-        return $this->merchantOrderId->get();
-    }
-
-    public function getExtra(): array
-    {
-        return $this->extra?->toArray();
-    }
-
-    public function getDescription(): string
-    {
-        return $this->description->get();
+        return $this->extra;
     }
 
     public function getCurrentTransaction(): Transaction