+++ src/Interfaces/MultiFieldsEntityInterface.php
@@ -7,4 +7,5 @@ interface MultiFieldsEntityInterface
     public function getField($fieldName): mixed;
     public function getPropertyName(): string;
     public function toArray(): array;
+    public function update(array $data, $index = null);
 }