+++ src/Interfaces/ValueInCentsInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace GingerPluginSdk\Interfaces;
-
-interface ValueInCentsInterface
-{
-    function calculateValueInCents($value);
-}