+++ CHANGELOG.md
@@ -90,3 +90,43 @@ Initial version
 * Updated `TransactionTest`.
 * Simplified `Event` entity.
 * Simplified `Transaction` entity.
+
+## 1.4.0
+
+* Implemented supporting `update` method for such collections :
+  * Order Lines.
+  * Abstract Collection.
+  * Additional Addresses.
+  * Transactions.
+  * Phone Numbers.
+* Implemented supporting `update` method for such entities : 
+  * Line.
+  * Address.
+  * Customer.
+  * Extra.
+  * Payment Method Details.
+  * Transaction.
+* Implemented `update` method to `MultiFieldEntityInterface`.
+* Implemented properties: 
+  * `RawCost`.
+  * `Percentage`.
+  * `VatPercentage`
+* Implemented tests: 
+  * `RawCostTest`.
+  * `AmountTest`.
+  * `VatPercentageTest`.
+  * `UpdateOrderTest`.
+* Updated and Simplified:
+  * `AdditionallAddressesTest`.
+  * `CustomerTtest`.
+  * `TransactionTest`.
+* Updated  `get` calls for such entities:
+  * Order.
+  * Customer.
+  * Transaction.
+* Redesigned `Amount` property to expect two possible variants in constructor `RawCost` property or value in cents.
+* Redesigned `fromArray` method to be static.
+* Updated tests to use `fromArray` as a static method.
+* Updated `fromArray` method.
+* Implemented `merchant_order_id` property for the `Order` entity.
+* Eliminated `ValueIncentsInterface`.